Move property details tab (READ + UPDATE)
