+++ source/frontend/src/features/mapSideBar/property/PropertyContainer.tsx
@@ -2,12 +2,12 @@ import { usePropertyDetails } from 'features/mapSideBar/hooks/usePropertyDetails
 import BcAssessmentTabView from 'features/mapSideBar/property/tabs/bcAssessment/BcAssessmentTabView';
 import LtsaTabView from 'features/mapSideBar/property/tabs/ltsa/LtsaTabView';
 import PropertyAssociationTabView from 'features/mapSideBar/property/tabs/propertyAssociations/PropertyAssociationTabView';
+import { PropertyDetailsTabView } from 'features/mapSideBar/property/tabs/propertyDetails/detail/PropertyDetailsTabView';
 import {
   InventoryTabNames,
   InventoryTabs,
   TabInventoryView,
 } from 'features/mapSideBar/tabs/InventoryTabs';
-import { PropertyDetailsTabView } from 'features/mapSideBar/tabs/propertyDetails/detail/PropertyDetailsTabView';
 import React, { useState } from 'react';
 
 import ComposedProperty from './ComposedProperty';
