+++ source/frontend/src/features/mapSideBar/property/PropertyViewSelector.tsx
@@ -1,4 +1,4 @@
-import { UpdatePropertyDetailsContainer } from 'features/mapSideBar/tabs/propertyDetails/update/UpdatePropertyDetailsContainer';
+import { UpdatePropertyDetailsContainer } from 'features/mapSideBar/property/tabs/propertyDetails/update/UpdatePropertyDetailsContainer';
 import { FormikProps } from 'formik';
 import * as React from 'react';
