+++ source/frontend/src/features/mapSideBar/property/tabs/propertyDetails/hooks/index.ts
