+++ source/frontend/src/features/mapSideBar/property/tabs/propertyDetails/hooks/useGetProperty.ts
