+++ source/frontend/src/features/mapSideBar/property/tabs/propertyDetails/hooks/useUpdateProperty.ts
