+++ source/frontend/src/features/mapSideBar/property/tabs/propertyDetails/update/UpdatePropertyDetailsContainer.test.tsx
