+++ source/frontend/src/features/mapSideBar/property/tabs/propertyDetails/update/models/PropertyAnomalyFormModel.ts
