+++ source/frontend/src/features/mapSideBar/property/tabs/propertyDetails/update/models/UpdatePropertyDetailsFormModel.ts
@@ -16,7 +16,7 @@ import {
   PropertyAnomalyFormModel,
   PropertyRoadFormModel,
   PropertyTenureFormModel,
-} from './';
+} from '.';
 
 export class AddressFormModel {
   id?: number;
