+++ source/frontend/src/features/mapSideBar/property/tabs/propertyDetails/update/styles.ts
