+++ source/frontend/src/features/mapSideBar/property/tabs/propertyDetails/update/validation.ts
