+++ source/frontend/src/features/properties/map/research/ViewSelector.tsx
@@ -1,10 +1,10 @@
+import { UpdatePropertyDetailsContainer } from 'features/mapSideBar/property/tabs/propertyDetails/update/UpdatePropertyDetailsContainer';
 import PropertyResearchTabView from 'features/mapSideBar/property/tabs/propertyResearch/PropertyResearchTabView';
 import {
   InventoryTabNames,
   InventoryTabs,
   TabInventoryView,
 } from 'features/mapSideBar/tabs/InventoryTabs';
-import { UpdatePropertyDetailsContainer } from 'features/mapSideBar/tabs/propertyDetails/update/UpdatePropertyDetailsContainer';
 import { FormikProps } from 'formik';
 import noop from 'lodash/noop';
 import { Api_ResearchFile } from 'models/api/ResearchFile';
